+++ spnkr/services/gamecms_hacs.py
@@ -6,6 +6,7 @@ from spnkr.models.gamecms_hacs import (
     MedalMetadata,
     SeasonCalendar,
 )
+from spnkr.responses import ImageResponse, JsonResponse
 from spnkr.services.base import BaseService
 
 _HOST = "https://gamecms-hacs.svc.halowaypoint.com"
@@ -14,16 +15,17 @@ _HOST = "https://gamecms-hacs.svc.halowaypoint.com"
 class GameCmsHacsService(BaseService):
     """Game content management data services."""
 
-    async def get_medal_metadata(self) -> MedalMetadata:
+    async def get_medal_metadata(self) -> JsonResponse[MedalMetadata]:
         """Get details for all medals obtainable in the game.
 
         Returns:
             The medal metadata.
         """
         url = f"{_HOST}/hi/Waypoint/file/medals/metadata.json"
-        return MedalMetadata(**await self._get_json(url))
+        resp = await self._get(url)
+        return JsonResponse(resp, lambda data: MedalMetadata(**data))
 
-    async def get_csr_season_calendar(self) -> CsrSeasonCalendar:
+    async def get_csr_season_calendar(self) -> JsonResponse[CsrSeasonCalendar]:
         """Get IDs and dates for past and current CSR seasons.
 
         CSR seasons represent the time periods between CSR resets.
@@ -34,9 +36,10 @@ class GameCmsHacsService(BaseService):
         url = (
             f"{_HOST}/hi/Progression/file/Csr/Calendars/CsrSeasonCalendar.json"
         )
-        return CsrSeasonCalendar(**await self._get_json(url))
+        resp = await self._get(url)
+        return JsonResponse(resp, lambda data: CsrSeasonCalendar(**data))
 
-    async def get_season_calendar(self) -> SeasonCalendar:
+    async def get_season_calendar(self) -> JsonResponse[SeasonCalendar]:
         """Get IDs and dates for past/current reward track events/operations.
 
         Reward tracks are the progression systems that allow players to earn
@@ -48,18 +51,20 @@ class GameCmsHacsService(BaseService):
         url = (
             f"{_HOST}/hi/progression/file/calendars/seasons/seasoncalendar.json"
         )
-        return SeasonCalendar(**await self._get_json(url))
+        resp = await self._get(url)
+        return JsonResponse(resp, lambda data: SeasonCalendar(**data))
 
-    async def get_career_reward_track(self) -> CareerRewardTrack:
+    async def get_career_reward_track(self) -> JsonResponse[CareerRewardTrack]:
         """Get details for the career rank progression reward track.
 
         Returns:
             The career rank reward track.
         """
         url = f"{_HOST}/hi/Progression/file/RewardTracks/CareerRanks/careerRank1.json"
-        return CareerRewardTrack(**await self._get_json(url))
+        resp = await self._get(url)
+        return JsonResponse(resp, lambda data: CareerRewardTrack(**data))
 
-    async def get_image(self, relative_path: str) -> bytes:
+    async def get_image(self, relative_path: str) -> ImageResponse:
         """Get an image from the game content management service.
 
         Args:
@@ -70,4 +75,4 @@ class GameCmsHacsService(BaseService):
             The image data.
         """
         url = f"{_HOST}/hi/images/file/{relative_path.lstrip('/')}"
-        return await self._get_bytes(url)
+        return ImageResponse(await self._get(url))