+++ spnkr/services/skill.py
@@ -5,6 +5,7 @@ from typing import Iterable
 from uuid import UUID
 
 from spnkr.models.skill import MatchSkill, PlaylistCsr
+from spnkr.responses import JsonResponse
 from spnkr.services.base import BaseService
 from spnkr.xuid import wrap_xuid
 
@@ -16,7 +17,7 @@ class SkillService(BaseService):
 
     async def get_match_skill(
         self, match_id: str | UUID, xuids: Iterable[str | int]
-    ) -> MatchSkill:
+    ) -> JsonResponse[MatchSkill]:
         """Get player CSR and team MMR values for a given match and player list.
 
         Args:
@@ -37,14 +38,15 @@ class SkillService(BaseService):
             raise ValueError("`xuids` cannot be empty")
         url = f"{_HOST}/hi/matches/{match_id}/skill"
         params = {"players": [wrap_xuid(x) for x in xuids]}
-        return MatchSkill(**await self._get_json(url, params=params))
+        resp = await self._get(url, params=params)
+        return JsonResponse(resp, lambda data: MatchSkill(**data))
 
     async def get_playlist_csr(
         self,
         playlist_id: str | UUID,
         xuids: Iterable[str | int],
         season_id: str | None = None,
-    ) -> PlaylistCsr:
+    ) -> JsonResponse[PlaylistCsr]:
         """Get player CSR values for a given playlist and player list.
 
         Args:
@@ -71,7 +73,8 @@ class SkillService(BaseService):
         params: dict = {"players": [wrap_xuid(x) for x in xuids]}
         if season_id:
             params["season"] = _clean_season_id(season_id)
-        return PlaylistCsr(**await self._get_json(url, params=params))
+        resp = await self._get(url, params=params)
+        return JsonResponse(resp, lambda data: PlaylistCsr(**data))
 
 
 def _clean_season_id(season_id: str) -> str: