+++ spnkr/services/stats.py
@@ -11,6 +11,7 @@ from spnkr.models.stats import (
     MatchStats,
     ServiceRecord,
 )
+from spnkr.responses import JsonResponse
 from spnkr.services.base import BaseService
 from spnkr.xuid import wrap_xuid_or_gamertag
 
@@ -29,7 +30,9 @@ _VALID_SERVICE_RECORD_FILTER_SETS = [
 class StatsService(BaseService):
     """Stats data services."""
 
-    async def get_match_count(self, player: str | int) -> MatchCount:
+    async def get_match_count(
+        self, player: str | int
+    ) -> JsonResponse[MatchCount]:
         """Get match counts across different game experiences for a player.
 
         The counts returned are for custom matches, matchmade matches, local
@@ -45,7 +48,8 @@ class StatsService(BaseService):
         """
         xuid_or_gamertag = wrap_xuid_or_gamertag(player)
         url = f"{_HOST}/hi/players/{xuid_or_gamertag}/matches/count"
-        return MatchCount(**await self._get_json(url))
+        resp = await self._get(url)
+        return JsonResponse(resp, lambda data: MatchCount(**data))
 
     async def get_service_record(
         self,
@@ -55,7 +59,7 @@ class StatsService(BaseService):
         game_variant_category: GameVariantCategory | int | None = None,
         is_ranked: bool | None = None,
         playlist_asset_id: str | UUID | None = None,
-    ) -> ServiceRecord:
+    ) -> JsonResponse[ServiceRecord]:
         """Get a service record for a player. Summarizes player stats.
 
         Note that filters (`season_id`, `game_variant_category`, `is_ranked`,
@@ -120,7 +124,8 @@ class StatsService(BaseService):
                 f"Invalid filter combination: {filters}. Options:\n{valid}"
             )
         params = {k.replace("_", ""): str(v) for k, v in filters.items()}
-        return ServiceRecord(**await self._get_json(url, params=params))
+        resp = await self._get(url, params=params)
+        return JsonResponse(resp, lambda data: ServiceRecord(**data))
 
     async def get_match_history(
         self,
@@ -128,7 +133,7 @@ class StatsService(BaseService):
         start: int = 0,
         count: int = 25,
         match_type: Literal["all", "matchmaking", "custom", "local"] = "all",
-    ) -> MatchHistory:
+    ) -> JsonResponse[MatchHistory]:
         """Request a batch of matches from a player's match history.
 
         Args:
@@ -147,9 +152,12 @@ class StatsService(BaseService):
         xuid_or_gamertag = wrap_xuid_or_gamertag(player)
         url = f"{_HOST}/hi/players/{xuid_or_gamertag}/matches"
         params = {"start": start, "count": count, "type": match_type}
-        return MatchHistory(**await self._get_json(url, params=params))
+        resp = await self._get(url, params=params)
+        return JsonResponse(resp, lambda data: MatchHistory(**data))
 
-    async def get_match_stats(self, match_id: str | UUID) -> MatchStats:
+    async def get_match_stats(
+        self, match_id: str | UUID
+    ) -> JsonResponse[MatchStats]:
         """Request match details using the Halo Infinite match GUID.
 
         Args:
@@ -159,4 +167,5 @@ class StatsService(BaseService):
             The match details.
         """
         url = f"{_HOST}/hi/matches/{match_id}/stats"
-        return MatchStats(**await self._get_json(url))
+        resp = await self._get(url)
+        return JsonResponse(resp, lambda data: MatchStats(**data))